+++ testcases.xml
@@ -1,0 +1,1 @@
+<testcases><testcase name="Create a tasklist&#10;"><steps><step><step_number>1</step_number><actions>Tap on plus button</actions><expectedresults>Create tasklist screen must be shown</expectedresults></step><step><step_number>2</step_number><actions>Tap on title label</actions><expectedresults>Keyboard must be shown</expectedresults></step><step><step_number>3</step_number><actions>Type the new tasklist name </actions><expectedresults>Label with the new tasklist name must be shown</expectedresults></step><step><step_number>4</step_number><actions>Tap on save button</actions><expectedresults>Tasklist must be shown</expectedresults></step></steps></testcase><testcase name="Create something&#10;"><steps><step><step_number>1</step_number><actions>action1</actions><expectedresults>expect1</expectedresults></step><step><step_number>2</step_number><actions>action2</actions><expectedresults>expect2</expectedresults></step><step><step_number>3</step_number><actions>action3</actions><expectedresults>expect3</expectedresults></step></steps></testcase><testcase name="Create something2&#10;"><steps><step><step_number>1</step_number><actions>action11</actions><expectedresults>expect11</expectedresults></step><step><step_number>2</step_number><actions>action22</actions><expectedresults>expect22</expectedresults></step><step><step_number>3</step_number><actions>action33</actions><expectedresults>expect33</expectedresults></step></steps></testcase><testcase name="Create something3&#10;"><steps><step><step_number>1</step_number><actions>action12</actions><expectedresults>expect12</expectedresults></step><step><step_number>2</step_number><actions>action23</actions><expectedresults>expect23</expectedresults></step><step><step_number>3</step_number><actions>action34</actions><expectedresults>expect34</expectedresults></step></steps></testcase></testcases>
