+++ testcaseswithoutsteps.xml
@@ -1,0 +1,2 @@
+<testcases>
+<testcase /><testcase name="Create a tasklist&#10;" /><testcase name="Create a task list with the same name that Tasks task list&#10;" /><testcase name="Create two tasks list with the same name &#10;" /><testcase name="Move a new created task to another existing task list&#10;" /><testcase name="Move a new created task to another new task list &#10;" /><testcase name="Create a task list after create a workspace&#10;" /><testcase name="Create a task list after a creation of a new user&#10;" /><testcase name="Create few tasks list in offline mode and check that those are created on web after turn on the connection" /></testcases>
